+++ goEmbeddedApi/api/models/data.go
@@ -9,7 +9,7 @@ type Data struct {
 	// Local     string    `json:"local"`
 	CreatedAt time.Time `gorm:"type:DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP" faker:"-"`
 	// TimeStamp time.Time `json:"time_stamp"`
-	TypeId int   `json:"type_id"`
-	Type   Types `gorm:"foreignKey:TypeId"`
-	Value  int   `json:"value"`
+	TypeId int     `json:"type_id"`
+	Type   Types   `gorm:"foreignKey:TypeId"`
+	Value  float64 `json:"value"`
 }
